Did a lot of work

+++ app/src/main/java/yegor/cheprasov/fluentflow/ui/compose/mainScreen/screens/themes/components/ThemeItem.kt
@@ -85,7 +85,7 @@ fun ThemeItem(
                     fontWeight = FontWeight.Light
                 )
                 Percentage(
-                    percentage = percentage,
+                    percentage = percentage.toFloat(),
                     modifier = Modifier
                         .width(134.dp)
                         .padding(top = 11.dp)
